requirements.txt
pyaes==1.6.1
Pyrogram==2.0.35
PySocks==1.7.1
python-dotenv==0.20.0
TgCrypto==1.2.3
